# website/docs/api/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Indexarr",
    "version": "0.1.0",
    "description": "Decentralized torrent indexing with DHT crawling, content classification, gossip-based P2P sync, and contributor identity.\n\n## Authentication\n\nMost endpoints require no authentication. Admin endpoints (API key management, nuke review, debug logging) require an API key passed via the `X-Api-Key` header. The Torznab endpoint supports an optional `apikey` query parameter.\n\n## Base URL\n\nAll REST endpoints are prefixed with `/api/v1/`. The Torznab endpoint is at `/api/torznab`. GraphQL is at `/graphql`."
  },
  "servers": [
    {
      "url": "http://localhost:8080",
      "description": "Local development"
    }
  ],
...
  "paths": {
...
    "/api/v1/torrent/{info_hash}/comments": {
      "get": {
        "tags": ["Social"],
        "summary": "List comments",
        "description": "Threaded comments for a torrent.",
        "parameters": [
          { "name": "info_hash", "in": "path", "required": true, "schema": { "type": "string" } }
        ],
        "responses": { "200": { "description": "Comment thread" } }
      },
      "post": {
        "tags": ["Social"],
        "summary": "Create comment",
        "parameters": [
          { "name": "info_hash", "in": "path", "required": true, "schema": { "type": "string" } }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "body": { "type": "string" },
                  "nickname": { "type": "string" },
                  "parent_id": { "type": "integer", "nullable": true }
                },
                "required": ["body"]
              }
            }
          }
        },
        "responses": { "200": { "description": "Comment created" } }
      }
    },
    "/api/v1/torrent/{info_hash}/comments/{comment_id}": {
      "delete": {
        "tags": ["Social"],
        "summary": "Delete comment",
        "description": "Soft-delete own comment (matched by fingerprint).",
        "parameters": [
          { "name": "info_hash", "in": "path", "required": true, "schema": { "type": "string" } },
          { "name": "comment_id", "in": "path", "required": true, "schema": { "type": "integer" } }
        ],
        "responses": { "200": { "description": "Comment deleted" } }
      }
    },
    "/api/v1/torrent/{info_hash}/votes": {
      "get": {
        "tags": ["Social"],
        "summary": "Vote summary",
        "description": "Upvotes, downvotes, and the current user's vote.",
        "parameters": [
          { "name": "info_hash", "in": "path", "required": true, "schema": { "type": "string" } }
        ],
        "responses": { "200": { "description": "Vote summary" } }
      }
    },
    "/api/v1/torrent/{info_hash}/vote": {
      "post": {
        "tags": ["Social"],
        "summary": "Cast vote",
        "parameters": [
          { "name": "info_hash", "in": "path", "required": true, "schema": { "type": "string" } }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "value": { "type": "integer", "enum": [1, -1] }
                },
                "required": ["value"]
              }
            }
          }
        },
        "responses": { "200": { "description": "Vote cast or toggled" } }
      }
    },
    "/api/v1/torrent/{info_hash}/nuke": {
      "post": {
        "tags": ["Social"],
        "summary": "Suggest nuke",
        "description": "Suggest a torrent for removal. One suggestion per fingerprint.",
        "parameters": [
          { "name": "info_hash", "in": "path", "required": true, "schema": { "type": "string" } }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "reason": { "type": "string" }
                },
                "required": ["reason"]
              }
            }
          }
        },
        "responses": { "200": { "description": "Nuke suggestion created" } }
      }
    },
    "/api/v1/nuke/pending": {
      "get": {
        "tags": ["Social"],
        "summary": "Pending nuke suggestions",
        "parameters": [
          { "name": "limit", "in": "query", "schema": { "type": "integer", "default": 50 } },
          { "name": "offset", "in": "query", "schema": { "type": "integer", "default": 0 } }
        ],
        "responses": { "200": { "description": "Unreviewed nuke suggestions grouped by torrent" } }
      }
    },
    "/api/v1/nuke/{suggestion_id}/review": {
      "post": {
        "tags": ["Social"],
        "summary": "Review nuke suggestion",
        "description": "Approve or dismiss a nuke suggestion. Requires API key.",
        "parameters": [
          { "name": "suggestion_id", "in": "path", "required": true, "schema": { "type": "integer" } }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "outcome": { "type": "string", "enum": ["nuked", "dismissed"] }
                },
                "required": ["outcome"]
              }
            }
          }
        },
        "responses": { "200": { "description": "Suggestion reviewed" } }
      }
    },
...
  }
}